pyproject.toml
[tool.poetry]
name = "zoo-store-back-end"
version = "0.1.0"
description = ""
authors = ["DmitryTok <FoB09@ukr.net>", "BRANYA43 <bogdan.timchenko.2504@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
django-allauth = "^0.58.2"
django-filter = "^23.4"
django-taggit = "^5.0.1"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
drf-spectacular = "^0.26.5"
rest-framework-simplejwt = "^0.0.2"
djangorestframework-simplejwt = "^5.3.0"
django-baton = "^2.8.1"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.5.0"
mypy = "^1.7.1"
django-debug-toolbar = "^4.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "migrations",
    "__init__.py",
]
line-length = 120

[tool.ruff.flake8-quotes]
docstring-quotes = "double"
inline-quotes = "single"
multiline-quotes = "single"